switch to a simpler thread pool design
In the previous thread pool design, tasks were FnOnce() + Send closures
which could be pushed onto an SPMC queue from the inside of a scope
closure. In this new design, tasks are Fn() + Sync instead, and the same
task runs identically on each of the threads in the thread pool.

Both the API and the implementation are simpler in this new design, and
it should still be completely adequate for parallelizing the renderer.