handle increments on the rightmost column in add_delta
Handle the case where a coverage increment falls on the rightmost column
of the canvas in add_delta, and don't pad the coverage buffer by one
pixel anymore.

This check does have a small performance penalty. The loop in add_line
should be restructured to amortize all of the checks in add_delta so
that they don't have to happen per-pixel.