don't pad coverage buffer width, and use load/store_partial
Don't pad coverage buffer width to a multiple of 4 (vector lane count).
Handle the case of a width not divisible by 4 when compositing by using
load_partial and store_partial.

This has a small performance penalty, and eventually should be handled
by treating the remainder at the end of a scanline separately.