rewrite bitscan loop to scan across multiple words
Rewrite the core bitscan loop in Rasterizer::finish() so that both the
scans for the next 1 bit and the scan for the next 0 bit search across
multiple words (previously, the former only searched within a single
word, and the latter only searched one bit at a time). This improves
performance for all vector widths, but especially for the scalar
backend.