replace (try_)specialize with (try_)invoke
In place of the try_specialize and specialize methods on PossibleArch
and SupportedArch, expose try_invoke and invoke methods which directly
run the given task. Additionally, add an invoke method to Arch, which
allows unconditionally invoking a task on the given architecture when
already inside another task running on that same architecture.