impl Task for FnOnce and use closures for Rasterizer methods
Rasterizer::add_line_inner and finish_inner don't need to perform
runtime feature dispatch, since that happens once in Rasterizer::new, so
all they actually need is to run some code in a #[target_feature]
context. Implementing Task for FnOnce makes it possible to wrap the body
of a method with

    A::invoke(#[inline(always)] || { ... })

to ensure that it is compiled with the appropriate target features
enabled.