separate PossibleArch::try_with and Arch::invoke
Previously, PossibleArch::try_invoke played two roles: first, it performed
runtime feature detection and invoked the given task with the
corresponding private Arch type if the required features were available,
and second, it invoked the given task in an #[target_feature] context.

Separate these two responsibilities so that PossibleArch::try_invoke
only performs feature detection and supplies the Arch parameter, and
then Arch::invoke takes a closure (not parameterized by Arch) and runs
it in a #[target_feature] context. Rename Task to WithArch and
(try_)invoke to (try_)with to make this more clear.