specialize add_segment for different segment directions
Define a trait, RasterConsts, with associated constants for the sign of
coverage deltas (whether the line segment is +Y or -Y) and the direction
of iteration (whether scanline traversal happens upwards or downwards).
Parameterize Rasterizer::add_segment with this trait and explicitly call
one of four specializations for a given segment.

This allows the optimizer to do more constant folding and results in an
improvement of 3-4% for the tiger benchmark on an Intel Core i5-8400.