replace RasterConsts with FlipCoords trait
Instead of a RasterConsts trait with associated constants, use a
FlipCoords trait to specialize add_segment for different segment
orientations. For segments which are traversed in the -Y direction,
mirror the Y-coordinates of the endpoints and then undo the mirroring
when computing row indices. This makes it possible to unify the two
halves of the branch on the value of C::Y_INCREMENT, as well as making
it so that both X and Y values are strictly increasing in the inner
loop.