make clipping logic more robust
Simplify the logic for clipping the y < 0 range of a segment, so that it
directly computes the clipped x coordinate rather than trying to step x by some
number of columns. Explicitly clip the y > width and x > height ranges of a
segment and remove the early-out logic from the inner rasterization loop.
Finally, when clipping the x < 0 range of a segment, step by entire scanlines
rather than single pixels. This last change fixes pathological behavior that
would occur for very long line segments directly to the left of the canvas.