rewrite stroking engine
Rewrite the stroking engine to be both more correct and more performant.
The previous stroking engine first fully flattened a path and then
computed offsets from the flattened path. Not only was this
fundamentally incorrect, as it erased the boundaries between curve
segments and made it impossible to properly handle stroke joins, it was
also inefficient, as both the flattened path and the stroked path were
allocated on the heap.

The new engine generates offset curves one segment at a time, using
tangent vectors to compute offset locations. Joins between segments are
now handled properly, although the only currently supported join style
is bevel (likewise, butt is still the only supported cap style).

The new engine also performs all stroke generation on the stack, feeding
output segments into a sink callabck without performing heap
allocations.

Stroking currently reuses the same flattening error tolerance logic that
is used for filled paths, which means that in situations with high
curvature and wide stroke widths, not enough segments are generated on
the convex side of the stroke and potentially too many are generated on
the concave side.  Additionally, degenerate cases such as cusps are not
yet handled in a fully general way.  Fixing both of these issues is left
to future work.