replace use of f32::floor with a faster handwritten version
On baseline x86_64, f32::floor ends up getting lowered to a function
call, and this handwritten version is significantly faster. This change
results in an 8.7% improvement in the tiger benchmark on the SSE2
backend. f32::floor gets lowered to a faster instruction sequence on
AVX2, so the improvement is much smaller there (on the order of 1%).