disable AVX2 backend and remove invoke! usage from rasterizer
After fixing the main performance issue with the SSE2 backend
(f32::floor being slower than on AVX2), the tiger benchmark is now only
about 5% slower than the AVX2 benchmark. Disable it for now since that
is not a big enough performance difference to justify the complexity of
runtime feature detection and #[target_feature] usage.