don't write bits past the end of the scanline
This fixes a bug introduced by the change to remove the 1px gutter on the
right-hand side of the coverage buffer, where a bit was being written into the
dirty bitset for the last coverage delta in a scanline unconditionally, even
when the corresponding pixel was out of bounds. It was rare for this bug to
have consequences, since writing extra bits into the dirty bitset can only ever
affect performance, not correctness, and this bug could only result in an
out-of-bounds panic if a path's bounding box filled the entire canvas and a
line segment intersected the bottom-right pixel.